requirements_gpu.txt
torch>=2.0.0
numpy>=1.21.0
keras>=2.12.0
Pillow>=9.0.0
tqdm>=4.65.0 